requirements.txt
django-ckeditor
django-rest-auth
